// src/app/core/services/model/functionsInput.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CirculationInput } from './circulationInput';
import { CirculationValveConfiguration } from './circulationValveConfiguration';
import { FlushScheduleInput } from './flushScheduleInput';
import { FunctionTypes } from './functionTypes';

export interface FunctionsInput { 
    /**
     * Is the function disabled?
     */
    disabledTimestamp?: string;
    flushSchedule?: FlushScheduleInput;
    /**
     * true = use legacy versions of flush and thermal disinfection
     */
    legacy?: boolean;
    /**
     * Name of the function.  Is unique, could be used as id
     */
    name: string;
    remark: string;
    type: FunctionTypes;
    circulationInput?: CirculationInput;
    /**
     * uuid of the key switch
     */
    keySwitchUuid?: string;
    /**
     * Uuid for the DTE (water heating element)
     */
    dteUuid?: string;
    /**
     * Uuid for UFC fault contact, only allowed for a circulation function
     */
    ufcFaultContactUuid?: string;
    /**
     * Use UFC, only allowed for a circulation function
     */
    useUFC?: boolean;
    /**
     * List of circulation valves
     */
    circulationValveConfigurationMapping?: { [key: string]: CirculationValveConfiguration; };
    /**
     * Device uuid's for the flush operation  The first element will be flushed first, the second element will be flushed second and so on.
     */
    flushableDevicePointUuids?: Array<string>;
    /**
     * Sensor for pwc
     */
    temperatureSensorPwcDevicePointUuid?: string;
}